Delete old Elevator subsystem, rename new one
Delete the old Elevator subsystem and replace the elevator package with the new one.
