+++ src/main/java/com/team1165/robot/subsystems/elevator/Elevator.java
@@ -7,191 +7,53 @@
 
 package com.team1165.robot.subsystems.elevator;
 
-import com.ctre.phoenix6.configs.Slot0Configs;
-import com.team1165.robot.globalconstants.FieldConstants.Reef;
-import com.team1165.robot.subsystems.elevator.constants.ElevatorConstants;
 import com.team1165.robot.subsystems.elevator.io.ElevatorIO;
-import com.team1165.robot.subsystems.elevator.io.ElevatorIOInputsAutoLogged;
+import com.team1165.robot.subsystems.elevator.io.ElevatorIO.ElevatorIOInputs;
 import com.team1165.robot.util.logging.LoggedTunableNumber;
-import edu.wpi.first.wpilibj.Alert;
-import edu.wpi.first.wpilibj.Alert.AlertType;
-import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import com.team1165.robot.util.statemachine.GoalOverridableStateMachine;
+import com.team1165.robot.util.statemachine.StateUtils;
+import edu.wpi.first.wpilibj.DriverStation;
+import java.util.EnumMap;
 import org.littletonrobotics.junction.Logger;
 
-public class Elevator extends SubsystemBase {
-  // Customizable PID values (if in tuning mode)
-  private static final LoggedTunableNumber kP =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kP", ElevatorConstants.gains.kP);
-  private static final LoggedTunableNumber kI =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kI", ElevatorConstants.gains.kI);
-  private static final LoggedTunableNumber kD =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kD", ElevatorConstants.gains.kD);
-  private static final LoggedTunableNumber kS =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kS", ElevatorConstants.gains.kS);
-  private static final LoggedTunableNumber kG =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kG", ElevatorConstants.gains.kG);
-  private static final LoggedTunableNumber kV =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kV", ElevatorConstants.gains.kV);
-  private static final LoggedTunableNumber kA =
-      new LoggedTunableNumber("Elevator/Tuning/Gains/kA", ElevatorConstants.gains.kA);
-  private static final LoggedTunableNumber positionOverride =
-      new LoggedTunableNumber("Elevator/Tuning/PositionOverride", 0.0);
-
+/** State-machine-based Elevator subsystem, powered by two motors. */
+public class Elevator extends GoalOverridableStateMachine<ElevatorState> {
   private final ElevatorIO io;
-  private final ElevatorIOInputsAutoLogged inputs = new ElevatorIOInputsAutoLogged();
+  private final ElevatorIOInputs inputs = new ElevatorIOInputs();
 
-  // Disconnected alerts
-  private final Alert leftDisconnected =
-      new Alert("Left ElevatorMotor disconnected!", AlertType.kError);
-  private final Alert rightDisconnected =
-      new Alert("Right ElevatorMotor disconnected!", AlertType.kError);
-
-  private double homePosition = 0.0;
-  private boolean emergencyStop = false;
-  private boolean brakeMode = false;
+  private final EnumMap<ElevatorState, LoggedTunableNumber> tunableMap =
+      StateUtils.createTunableNumberMap(name + "/Positions", ElevatorState.class);
 
   public Elevator(ElevatorIO io) {
+    super(ElevatorState.IDLE);
     this.io = io;
   }
 
   @Override
-  public void periodic() {
+  protected void update() {
     io.updateInputs(inputs);
-    Logger.processInputs("Elevator", inputs);
-
-    LoggedTunableNumber.ifChanged(
-        hashCode(),
-        () ->
-            io.setPID(
-                new Slot0Configs()
-                    .withKP(kP.get())
-                    .withKI(kI.get())
-                    .withKD(kD.get())
-                    .withKG(kG.get())
-                    .withKS(kS.get())
-                    .withKV(kV.get())
-                    .withKA(kA.get())),
-        kP,
-        kI,
-        kD,
-        kG,
-        kS,
-        kV,
-        kA);
-
-    // Set alerts
-    leftDisconnected.set(!inputs.leftMotorConnected);
-    rightDisconnected.set(!inputs.rightMotorConnected);
+    Logger.processInputs(name, inputs);
   }
 
-  public void stop() {
-    io.stop();
-    if (!emergencyStop) {
-      Logger.recordOutput("Elevator/RunningMode", "Stop");
+  @Override
+  protected void transition() {
+    switch (getCurrentState()) {
+      case STOP -> io.stop();
+      case ADAPTIVE_L1, ADAPTIVE_L2, ADAPTIVE_L3, ADAPTIVE_L4 -> {
+        // TODO: Add adaptive code
+        break;
+      }
+      default -> {
+        if (tunableMap.containsKey(getCurrentState())) {
+          io.runPosition(tunableMap.get(getCurrentState()).get());
+        } else {
+          DriverStation.reportError(
+              "Elevator state "
+                  + getCurrentState()
+                  + " has no provided value or alternatives in the transition() method! Continuing without setting the state.",
+              false);
+        }
+      }
     }
-  }
-
-  public void runAtPercent(double percent) {
-    if (!emergencyStop) {
-      Logger.recordOutput("Elevator/RunningMode", "BasicVoltage");
-      io.runVolts(percent * 12.0);
-    }
-  }
-
-  public void runVolts(double volts) {
-    if (!emergencyStop) {
-      Logger.recordOutput("Elevator/RunningMode", "BasicVoltage");
-      io.runVolts(volts);
-    }
-  }
-
-  public void runPosition(double positionInches) {
-    if (!emergencyStop) {
-      Logger.recordOutput("Elevator/Setpoint", positionInches);
-      Logger.recordOutput("Elevator/RunningMode", "ManualPosition");
-      io.runPosition(positionInches + homePosition);
-    }
-  }
-
-  public void runPosition(Reef.Level level) {
-    if (!emergencyStop) {
-      Logger.recordOutput("Elevator/Setpoint", level.getElevatorHeight());
-      Logger.recordOutput("Elevator/RunningMode", "Coral " + level.name());
-      io.runPosition(level.getElevatorHeight() + homePosition);
-    }
-  }
-
-  public void runToIntakePosition() {
-    if (!emergencyStop) {
-      Logger.recordOutput("Elevator/Setpoint", ElevatorConstants.intakePosition);
-      Logger.recordOutput("Elevator/RunningMode", "Intake");
-      io.runPosition(ElevatorConstants.intakePosition + homePosition);
-    }
-  }
-
-  public Command runToNextPosition() {
-    return startRun(
-        () -> {
-          if (!emergencyStop) {
-            if (Math.abs(ElevatorConstants.intakePosition - inputs.leftPositionInches) > 0.2) {
-              io.runPosition(Reef.Level.L1.getElevatorHeight());
-            } else if (Math.abs(Reef.Level.L1.getElevatorHeight() - inputs.leftPositionInches)
-                > 0.2) {
-              io.runPosition(Reef.Level.L2.getElevatorHeight());
-            } else if (Math.abs(Reef.Level.L2.getElevatorHeight() - inputs.leftPositionInches)
-                > 0.2) {
-              io.runPosition(Reef.Level.L3.getElevatorHeight());
-            } else if (Math.abs(Reef.Level.L3.getElevatorHeight() - inputs.leftPositionInches)
-                > 0.2) {
-              io.runPosition(Reef.Level.L4.getElevatorHeight());
-            } else {
-              io.runPosition(ElevatorConstants.intakePosition);
-            }
-          }
-        },
-        () -> {});
-  }
-
-  public void setBrakeMode(boolean enabled) {
-    if (!emergencyStop) {
-      brakeMode = enabled;
-      io.setBrakeMode(enabled);
-    }
-  }
-
-  /**
-   * Emergency stop the elevator, immediately stop motors and set to brake mode. Run methods will
-   * not do anything.
-   */
-  public void setEmergencyStop(boolean enabled) {
-    if (enabled) {
-      Logger.recordOutput("Elevator/RunningMode", "EmergencyStop");
-      emergencyStop = true;
-      io.stop();
-      io.setBrakeMode(false);
-    } else {
-      Logger.recordOutput("Elevator/RunningMode", "Stop");
-      emergencyStop = false;
-      io.setBrakeMode(brakeMode);
-    }
-  }
-
-  public void setHomePosition(double homePositionInches) {
-    homePosition = homePositionInches;
-  }
-
-  public boolean getAtPosition(double positionInches, double tolerance) {
-    var currentPosition = inputs.leftPositionInches + homePosition;
-    return (positionInches - tolerance < currentPosition)
-        && (positionInches + tolerance > currentPosition);
-  }
-
-  public double getPosition() {
-    return inputs.leftPositionInches;
-  }
-
-  public boolean getEmergencyStopState() {
-    return emergencyStop;
   }
 }
