+++ src/main/java/com/team1165/robot/subsystems/elevator/ElevatorConstants.java
@@ -5,7 +5,7 @@
  * the root directory of this project.
  */
 
-package com.team1165.robot.subsystems.newelevator;
+package com.team1165.robot.subsystems.elevator;
 
 import com.ctre.phoenix6.configs.MotionMagicConfigs;
 import com.ctre.phoenix6.configs.Slot0Configs;
